refactor(middleware): organize middleware into several smaller files
