+++ src/main/ipcs/clipboard.ts
@@ -2,7 +2,7 @@ import { ipcMain, dialog, BrowserWindow, shell, clipboard } from 'electron'
 import { PLATFORM, ENVIRONMENT } from 'shared/constants'
 
 // helper functions
-const { IPC_EVENT_copyToClipboard } = require('../shared/main-renderer-events')
+const { IPC_EVENT_copyToClipboard } = require('../../shared/main-renderer-events')
 
 const copyToClipboard = (str) => {
     clipboard.writeText(str)
