+++ src/main/menu.ts
@@ -2,7 +2,7 @@ import { BrowserWindow, dialog, MenuItem } from 'electron'
 import { selectStatus } from 'shared/data/slices/pipeline'
 import { calculateJobName, readableStatus } from 'shared/jobName'
 import { Job, JobState, JobStatus, PipelineStatus } from 'shared/types'
-import { getPipelineInstance } from './data/middlewares/pipeline'
+import { getPipelineInstance } from './data/instance'
 import { store } from './data/store'
 import { closeApplication } from './windows'
 import { selectEditOnNewTab } from 'shared/data/slices/settings'
